"Simple client-server framework
Sets up a connection between the front-end user interface (HTML, Javascript, CSS) and back-end classification (Python, Pytorch, Torchvision) using the Flask micro-framework. A user can upload a single image, which is sent to classify.py through an Ajax call and Flask routing. The selected pre-trained Torchvision model classifies the image, returns the label/error and file name, and the results and image are displayed on the webpage.

See README.md for more details and next steps."